Initial support for elasticsearch added
The docker-compose.yml file was extended to start a rabbitmq service (for celery) as well as a elasticsearch node.
Dependencies to elasticsearch-dsl/elasticsearch for querying and indexing and a new  command line function (setup_es)  for indexing genes and SNPs in ES were added.
To make it easer to get started a shell which will take care of downloading the GFF3 and snpeff file was added as well.
Additionally some system checks were added that are run when the backend is started to make sure that elasticsearch is initialized.