Added GeneSearch component
The GeneSearch component uses vuetify's select component to display an autocomplete list.
The component is now used instead of a normal text field inside the geneDetail view. When a gene is selected int he GeneSearch component the geneDetail view will update the url and the information is updated.

Changed the breadcrumbs to be a computed property that depends on the selectedGene.