Implemented indexing of top associations
Added a command line function to index studies.
Users can provide a id to index a specific study. If no id is provided
all studies will be indexed. The indexing function itself is implemented
as a celery task but the command line calls it direcctly.
Additionally the get_top_association function was modified to be more
efficient.