Added proper support for intro.js
Removed the directives and replaced it with a mixin (Tour), that can be added to any view that should support a tour.
In order for a view to participate in a tour, you need to define a tourOptions object that contains the steps. If the tour should continue to another view, also sepcify a nextPage property that is passed directly to the router. This way it is possible to provide parameters to the route for the next page.

The mixin will check the route for a ?tour query parameter and start the tour if it is available.
Furhtermore Google Analytics events are sent when a tour starts, completes or is aborted.