Fix tests and implementing optional filtering by maf
By default the get_top_associations function will apply a maf filter of 0.05. This can be overriden. To return all top assocations just pass maf=0.
The implementation tries to be as efficient as possible and avoids work where possible.

In case of a threshold filter it will filter the SNPs.
In case of top number it will fetch and filter by maf as long as the requested amount is retrieved or the end of the pvalue list is reached.
Also added additional tests to make sure that all combinations work.