Bugfix: Edge case for drawing manhattan plots
If the gene view is opened the first time, it can happen that the region is set after the first draw, which can cause issues because th eregion is set to [-5000: 5000] which draws the points outside of the viewable region.
Fix is to make sure to call draw when the region is set. Might issue redudnant draw calls but d3js should be smart enough to only update things that have changed.