Optimize performance in gene detail view
Instead of re-rendering all associations in the manhattan plot when highlighting a list of associations, re-render only the highlighted associations. For this the list of highlighted associations has to be cached in the manhattanplot.js.
When highlighting a different set the previous highlighted associations are re-rendered followed by the new list.

Additionally avoiding redundant calls to the highlightassociations function.
When highlighting an associations now only the size and fill-opacity is changed, which should be more performant.
Renamed events from highlightsnp/unhighlightsnp to highlightassociation/unhighlightassociation.
This fixes #70