Add ability to specify the coloring of associations
The associations in the Manhattan plot can be colored based on some specific properties. The colorlegend component displays the coloring of the associations. There are two types of visualizations. For quantitive properites a sequential color band is displayed. For non-numeric properties, a categorical legend is displayed.
Currently following proprties are supported (Same color, maf, mac, impact and score).

The colorlegend is interactive. When the user hoves over an associations the corresponding color legend is highlighted. In case of categorical properties the user can also hover a category and the corr. associations in the Manhattan plot are highlighted.

This closes #59 and #60