Upgraded vuetify, vue.js and typescript dependencies
Vuetify was upgraded from 0.13.1 to 0.14.1 and Vue.js to 2.4 and typescript to 2.4.1
This causes some issues:
Vuetify deprecated some components (v-card-row, etc) and the v-data-table defines the headers now differently.
Typescript is a bit more strict with type inference, so the types had to be specified in the api calls.

Additionally highlighted associations are now properly highlighted also in the association table (#57)