New Manhattan plots with hover information
Completely redesigned manhattan vue component and adapted manhattanplot.js file to have well-functioning hover functionality. Reduced number of displayed SNPs to 3000 for faster rendering. The js code has a different structure than the one used in the gene plot because the data is stored as an array of values and not as an array of Association Objects (to make it faster). All transitions were also avoided to limit the browser work. Solves #77