Fix: Use Suggest completion when searching for a gene name/id
Instead of using a normal match against _all use the completion suggester for searching for a gene name or id. This will properly displa results for partial query terms.
Additionally when searching for a chr and a start position, it will not display all the genes that are upstream of that start position instead of an empty list.

This fixes #80