Added pre-loaded json file for heatmap
A pre-computed json file link was added to speed up loading time. The number of hits per study was slightly increased (40) and the point size augmented. Adding any argument to the ?recompute parameter of the rest query recompute the data from scratches. The default region_width for heatmap retrieval was also increased to allow for faster recomputation (~15s).