Download funcitonalities
Implemented the associations download functionalities. the buttons are located within the filters bars and allow for download of filteres associations. A CSV file is generated in the backend and a periodical celery task is expected to clean all temporary files. If the whole dataset is selected, a pre-computed file is delivered. Closes #44.