Split up django settings into prod and dev files
Instead of using one settings.py file, we split up the settings into a prod.py and dev.py. By default the dev.py is called by the __init__.py in the settings module.
This can be overridden by setting the DJANGO_SETTINGS_MODULE to aragaws.settings.prod which is done in the docker-compose.yml file.

Additionally for production DEBUG is set to FALSE and errors will be reported by emails to the ADMINS