Optimize top association component for mobile
Refactor topAssoc.vue component to only have one instance of the table
and filter controls and use flexbox to position it either on the left or
right. Additionally automatically hide the filter controls when the
available width is less then 1092px. Fix the width of the filter
controls to 360px. This closes #105