Add study_name to phenotype model and fix celery
Add the name of the study from Arapheno to the Phenotype model.
This helps to distinguish phenotypes/GWAS studies with the same name (i.e. FTP10).
Additionally upgrade celery version to avoid syntax error
in Python 3.7 ("async" is reserved keyword).
Remove the 15 minutes trigger from Jenkinsfile as we are using webhooks.